Backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.2
ecdsa==0.19.1
email-validator>=2.0.0
exceptiongroup==1.3.0
fastapi>=0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg2-binary>=2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic>=2.10.0
pydantic-settings==2.1.0
pydantic_core>=2.27.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy>=2.0.36
starlette>=0.40.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
